Title: 스승의 은혜 2, Time: 118ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240922/스승의 은혜 2/the-grace-form-teacher-2.py
@@ -13,7 +13,7 @@ for i in range(n):
     for j in range(n):
         cnt += 1
         if j == i:
-            temp_sum += arr[j] / 2;
+            temp_sum += int(arr[j] / 2);
         else: 
             temp_sum += arr[j]
